A: ad-shield
Ad Shield is an ad recovery solution. They are utilizing multiple domains in an attempt to get around blocking. The two we have seen so far are css-load.com and 07c225f3.online, both of which return a generic message:

```This domain is used to load critical website contents such as HTML, CSS, image, video, etc.

If this domain is blocked, websites using the domain to load critical elements may not load / function properly.

To any relevant applications / anti-virus software / browser extensions / adblockers: please make sure to not place this domain on the blacklists.```